+++ src/main/java/cn/haitaoss/tinyioc/beans/annotation/Autowired.java
@@ -1,0 +1,18 @@
+package cn.haitaoss.tinyioc.beans.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author haitao.chen
+ * @email haitaoss@aliyun.com
+ * @date 2021-04-22 21:51
+ *
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD, ElementType.FIELD})
+public @interface Autowired {
+    String getId() default "";
+}