+++ src/main/java/cn/haitaoss/tinyioc/beans/annotation/Controller.java
@@ -1,0 +1,18 @@
+package cn.haitaoss.tinyioc.beans.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * @author haitao.chen
+ * @email haitaoss@aliyun.com
+ * @date 2021-04-23 10:23
+ *
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Controller {
+    String value() default "";
+}