+++ src/main/java/cn/haitaoss/tinyioc/exception/AnnotationException.java
@@ -1,0 +1,13 @@
+package cn.haitaoss.tinyioc.exception;
+
+/**
+ * @author haitao.chen
+ * @email haitaoss@aliyun.com
+ * @date 2021-04-22 21:57
+ *
+ */
+public class AnnotationException extends RuntimeException {
+    public AnnotationException(String message) {
+        super(message);
+    }
+}