+++ changelog.md
@@ -52,13 +52,62 @@ public Object invoke(Object proxy, Method method, Object[] args) throws Throwabl
 # 优化代理
 - 优化代码，如果被代理对象有实现接口，就使用jdk代码。否则使用cglib代理
 - 向三级缓存插入数据的位置调整
-- 
 
 问题：
-    - 使用cglib进行代理，只能通过set方法注入属性。field.set() 设置的是代理对象的属性不是被代理对象的
-    - 使用jdk进行代理，无法通过反射设置值。method.invoke() filed.set() 都出错。
+
+    -  使用cglib进行代理，只能通过set方法注入属性。field.set() 设置的是代理对象的属性不是被代理对象的 
+    -  使用jdk进行代理，无法通过反射设置值。method.invoke() filed.set() 都出错。
 
 # 实现构造器的解析
+
+```java
+protected Object doCreateBean(String name, BeanDefinition beanDefinition) throws Exception {
+        Object bean = createBeanInstance(beanDefinition);
+        // thirdCache中放置的全是空构造方法构造出的实例
+        thirdCache.put(name, bean);
+        beanDefinition.setBean(bean);
+        applyPropertyValues(name, bean, beanDefinition);
+        // 解析类里面的注解
+        injectAnnotation(name, bean, beanDefinition);
+        // 生命周期
+        if (bean instanceof InitializingBean) {
+            ((InitializingBean) bean).afterPropertiesSet();
+        }
+        return bean;
+    }
+```
+
+```java
+private Object createBeanInstance(BeanDefinition beanDefinition) throws Exception {
+        if (beanDefinition.getConstructorArgument().isEmpty()) {
+            return beanDefinition.getBeanClass().newInstance();
+        } else {
+            List<ConstructorArgument.ValueHolder> valueHolders = beanDefinition.getConstructorArgument().getArgumentValues();
+            Class clazz = Class.forName(beanDefinition.getBeanClassName());
+            Constructor[] cons = clazz.getConstructors();
+            for (Constructor constructor : cons) {
+                // 这里省去判断参数类型
+                if (constructor.getParameterCount() == valueHolders.size()) {
+                    Object[] params = new Object[valueHolders.size()];
+                    for (int i = 0; i < params.length; i++) {
+                        params[i] = valueHolders.get(i).getValue();
+                        if (params[i] instanceof BeanReference) {
+                            BeanReference ref = (BeanReference) params[i];
+                            String refName = ref.getName();
+                            if (thirdCache.containsKey(refName) && !firstCache.containsKey(refName)) {
+                                throw new IllegalAccessException("构造函数循环依赖" + refName);
+                            } else {
+                                params[i] = getBean(refName);
+                            }
+                        }
+                    }
+                    return constructor.newInstance(params);
+                }
+            }
+        }
+        return null;
+    }
+```
 
 # 增加类型解析器，解析复杂类型的属性
 - 我们自定义的类型的属性的赋值。我们可能有一些定制化需求。所以我们可以通过向容器中加入Converter来实现。Converter接口的parse方法的返回值，就是最终注入个属性的值
@@ -103,13 +152,136 @@ if (bean == null) {
 }
 ```
 - doCreateBean：给bean的属性赋值之后会调用bean的afterPropertiesSet 初始化bean
-  ```java
+```java
 applyPropertyValues(name, bean, beanDefinition);
 // 生命周期
 if(bean instanceof InitializingBean){
     ((InitializingBean) bean).afterPropertiesSet();
 }
-    ```
+```
 
 - initializeBean：调用BeanPostProcessor 的before方法后会获取bean的 init_method 初始化
 - applicationContext.close()：调用bean的销毁方法
+
+# 实现扫描注解加载bean
+
+## 解析类里面的注解，注入属性
+
+```java
+ protected Object doCreateBean(String name, BeanDefinition beanDefinition) throws Exception {
+        Object bean = createBeanInstance(beanDefinition);
+        // thirdCache中放置的全是空构造方法构造出的实例
+        thirdCache.put(name, bean);
+        beanDefinition.setBean(bean);
+        applyPropertyValues(name, bean, beanDefinition);
+   		
+        // 解析类里面的注解
+        injectAnnotation(bean, beanDefinition);
+
+        // 生命周期
+        if (bean instanceof InitializingBean) {
+            ((InitializingBean) bean).afterPropertiesSet();
+        }
+        return bean;
+    }
+```
+## 设置单例bean的标记
+
+```java
+private void processBeanDefinition(Element ele) {
+        String name = ele.getAttribute("id");
+        String className = ele.getAttribute("class");
+        BeanDefinition beanDefinition = new BeanDefinition();
+        // 解析是否是单例
+        String scope = ele.getAttribute("scope");
+        if (scope == null || scope.length() == 0 || scope.equals("singleton")) {
+            beanDefinition.setSingleton(true);
+        } else {
+            beanDefinition.setSingleton(false);
+        }
+        // 解析构造器参数
+        processConstructorArgument(ele, beanDefinition);
+        processProperty(ele, beanDefinition);
+        beanDefinition.setBeanClassName(className);
+        getRegistry().put(name, beanDefinition);
+
+    }
+```
+## 判断是不是单例bean
+
+```java
+public Object getBean(String name) throws Exception {
+        BeanDefinition beanDefinition = beanDefinitionMap.get(name);
+        if (beanDefinition == null) {
+            throw new IllegalArgumentException("No bean named " + name + "is defined");
+        }
+        Object bean = beanDefinition.getBean(); // null
+        // 如果bean为null 或者不是单例bean
+        if (bean == null || !beanDefinition.isSingleton()) {
+            bean = doCreateBean(name, beanDefinition);
+            bean = initializeBean(bean, name); // 代理操作
+            // 将操作过的bean重新设置到beanDefinition中
+            beanDefinition.setBean(bean); // 修改beandefinition 里面的bean
+        }
+        return bean;
+    }
+```
+## AnnotationParser
+
+1. 根据packageName 收集当前包和子包下面的所有非内部类(Set<String>)
+
+2. 遍历Set，Class.forName(className).isAnnotation(xx)。判断当前class 是否有Component、Repository、Service、Controller 注解
+
+3. 存在注解，那就创建BeanDefinition。然后将生成BeanDefinition 添加到map中临时存储
+    - 获取BeanDefinition的name值。annotation.value(); 获取不到就取当前类(首字母小写)作为name的值
+    - 获取当前class 里面所有的属性，判断是否标注了@Value注解，如果有就创建成PropertyValue，添加到BeanDefinition的PropertyValues中
+
+4. 使用AnnotationParser
+
+    - 修改loadBeanDefinitions 方法的内容
+
+      ```java
+       public void refresh() throws Exception {
+              // 将读取xml 获取的容器复制到 beanFactory中
+              loadBeanDefinitions(beanFactory);
+              // 注册类型转换器
+              registerConverter(beanFactory);
+              // 注册beanPostProcessor
+              registerBeanPostProcessors(beanFactory);
+              // 创建出ioc容器中所有的对象
+              onRefresh();
+          }
+      ```
+
+      
+
+    - 不仅仅会注册读取xml得到的BeanDefinition，而且也要注册通过扫描注解生成的BeanDefinition
+
+    ```java
+        @Override
+        public void loadBeanDefinitions(AbstractBeanFactory beanFactory) throws Exception {
+            // 去取xml配置文件注册 BeanDefinition
+            XmlBeanDefinitionReader xmlBeanDefinitionReader = new XmlBeanDefinitionReader(new ResourceLoader());
+            xmlBeanDefinitionReader.loadBeanDefinitions(configLocation);
+    
+            for (Map.Entry<String, BeanDefinition> entry : xmlBeanDefinitionReader.getRegistry().entrySet()) {
+                beanFactory.registerBeanDefinition(entry.getKey(), entry.getValue());
+            }
+    
+            // 扫描主包下面的类注册 BeanDefinition
+            String packageName = xmlBeanDefinitionReader.getPackageName();
+            if (packageName == null) {
+                return;
+            }
+            AnnotationParser annotationParser = new AnnotationParser();
+            annotationParser.annotationBeanDefinitionReader(packageName);
+            for (Map.Entry<String, BeanDefinition> entry : annotationParser.getRegistry().entrySet()) {
+                beanFactory.registerBeanDefinition(entry.getKey(), entry.getValue());
+            }
+        }
+    
+    
+    ```
+
+    
+